requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
SQLAlchemy==2.0.41
supabase==2.18.1
python-dotenv==1.1.1
psycopg[binary,pool]==3.2.10
Pillow==11.2.1
pywebpush==2.0.3